MainController __init__ now takes a path to the configuration directory instead of specifying the path to each config. MainController is meant to be used with a configuration directory, hence the switch.
This change allows all test configs to be opened through settings when
running test_blade_runner_manual.py, which wasn't the case previously.

Updated test_blade_runner_manual.py to be compatible with these changes.

Added _reload_configs(). Reloads configuration files in the
configuration directory.

MainView has been updated to call _reload_configs() when the settings
scene is left. This essentially ensures that the configs are reloaded
if any changes are made. Before this commit, the only way to reload
the configs was to quit Blade Runner and start it up again.